+++ src/main/java/com/flab/tabling/user/domain/Member.java
@@ -1,46 +1,46 @@
 package com.flab.tabling.user.domain;
 
 import com.flab.tabling.global.util.BaseTime;
-import com.flab.tabling.user.util.RoleTypeConverter;
-import jakarta.persistence.Column;
-import jakarta.persistence.Convert;
+
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ * @Entity: 엔티티 지정, 객체와 테이블 매핑
+ * @NoArgsConstructor(access = AccessLevel.PROTECTED) : 기본 생성자를 protected 접근 제어자로 설정
+ * 무분별한 기본 생성자 제한을 막기 위한 것.
+ * @Id: 기본키 지정
+ * @GeneratedValue(strategy = GenerationType.IDENTITY) : 대리 키 사용 방식,
+ * IDENTITY: 기본 키 생성을 데이터베이스에 위임
+ * @Builder : 빌더 패턴
+ */
+
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Member extends BaseTime {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @NotNull
-    @Size(max = 30)
-    private String name;
-    @NotNull
-    @Size(max = 255)
-    private String email;
-    @NotNull
-    @Size(max = 255)
-    private String password;
-    @NotNull
-    @Column(length = 20)
-    @Convert(converter = RoleTypeConverter.class)
-    private RoleType roleType;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	private String name;
+	private String email;
+	private String password;
+	@Enumerated(EnumType.STRING)
+	private RoleType roleType;
 
-    @Builder
-    public Member(String name, String email, String password, RoleType roleType) {
-        this.name = name;
-        this.email = email;
-        this.password = password;
-        this.roleType = roleType;
-    }
+	@Builder
+	public Member(String name, String email, String password, RoleType roleType) {
+		this.name = name;
+		this.email = email;
+		this.password = password;
+		this.roleType = roleType;
+	}
 }
