+++ src/main/java/com/flab/tabling/user/exception/MemberExistException.java
@@ -1,0 +1,4 @@
+package com.flab.tabling.user.exception;
+
+public class MemberExistException extends RuntimeException {
+}
