--- src/main/java/com/flab/tabling/user/exception/UserExistException.java
@@ -1,4 +1,0 @@
-package com.flab.tabling.user.exception;
-
-public class UserExistException extends RuntimeException{
-}
